fix: update the notification receiver to wait until the full message is received before returning
When the MTU of the BLE connection is less than the size of the
notification message (30 bytes in this case), an Airthings Wave+ (and
possibly other models, but I don't have them to verify) will send the
message over multiple packets. This PR teaches the code to wait for the
full message.