Max max attempts adjustable
The default is set to one attempt so by default it
only tries once. The design is to let it fail at
startup and have setup retry kick in so it tries
again after startup and does not delay startup
when its have a bit of trouble. Once startup
has completed, the max_attempts can be set to
a higher value since it will not be delaying
startup.