Add VOC level strings sensor
This is similar to the Radon level strings, but for VOC.

The levels and descriptions were taken from [1]. I merely copied the
ranges provided by Airthings and do not claim that they are reasonable,
accurate or otherwise.

This is convenient as an end-user has no idea what are acceptable
measurements, giving them a text sensor makes it easier to spot
problems.

[1]: https://www.airthings.com/en-ca/contaminants/what-is-voc